Implemented uniform locations caching
For now it's pretty dumb with linear comparison. Probably should be replaced with associative array, but feels like we have just a few items in it and in general it's going to be an overkill.

For now it should fix the issue with very slow uniform location lookup by nvidia

Feels like it doesn't help much or i'm interpreting profiler in a wrong way